servicios terminados

+++ Visor/src/app/components/configuraciones/permisos-roles/permisos-roles.component.ts
@@ -33,10 +33,11 @@ export class PermisosRolesComponent implements OnInit {
     this.actualizarPermisos();
     this.ordenAlfabetico(this.permisos);
     this.dataSource = new MatTableDataSource(this.permisos);
-    this._usuarioService.verificarUsuario('F001', localStorage.getItem('usuario')!).subscribe(response => {
-      this.privilegio = response.returnset.Rcode; // agregar el if
+    this._usuarioService.verificarUsuario('F001', localStorage.getItem('usuario')!).subscribe((response: any) => {
+      if (response.returnset[0].RCode == 200) {
+        this.privilegio = true;
+      }
     });
-    console.log(this.privilegio)
   }
 
   getRoles() {
@@ -84,6 +85,7 @@ export class PermisosRolesComponent implements OnInit {
   }
 
   marcarCheckbox(event: any, rol: string, funcion: string) {
+    console.log(this.privilegio)
     this.dialogService.openConfirmDialog(
       '¿Está seguro de que desea cambiar el permiso de este rol?'
       ).afterClosed().subscribe(res => {
@@ -164,17 +166,29 @@ export class PermisosRolesComponent implements OnInit {
               });
               if (event.checked) {
                 console.log("insert", id_rol, id_funcion);
-                this._permisoService.insertPermiso(id_funcion, id_rol).subscribe((response: any) => {
-                  console.log("INSERT EXITOSO", response);
+                if (this.privilegio == true) {
+                  this._permisoService.insertPermiso(id_funcion, id_rol).subscribe((response: any) => {
+                    console.log("INSERT EXITOSO", response);
                   });
+                  permiso.supervisor = event.checked;
+                  this.mensajeCambio('El permiso para el rol ' + rol + ' fue modificado correctamente');
+                } else {
+                  permiso.supervisor = !event.checked;
+                  this.mensajeCambio('Su usuario no tiene los privilegios para cambiar permisos');
+                }
               } else {
                 console.log("delete", id_funcion_rol);
-                this._permisoService.deletePermiso(id_funcion_rol).subscribe((response: any) => {
-                  console.log("DELETE EXITOSO", response);
-                });
+                if (this.privilegio == true) {
+                  this._permisoService.deletePermiso(id_funcion_rol).subscribe((response: any) => {
+                    console.log("DELETE EXITOSO", response);
+                  });
+                  permiso.supervisor = event.checked;
+                  this.mensajeCambio('El permiso para el rol ' + rol + ' fue modificado correctamente');
+                } else {
+                  permiso.supervisor = !event.checked;
+                  this.mensajeCambio('Su usuario no tiene los privilegios para cambiar permisos');
+                }
               }
-              permiso.supervisor = event.checked;
-              this.mensajeCambio('El permiso para el rol ' + rol + ' fue modificado correctamente');
             } else {
               permiso.supervisor = !event.checked;
             }
